test(tooltip): unskip tests
